+++ src/main/java/net/gym/login/controller/LoginController.java
@@ -1,0 +1,61 @@
+package net.gym.login.controller;
+
+import java.io.IOException;
+
+import javax.servlet.RequestDispatcher;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+import net.gym.login.dao.LoginDao;
+import net.gym.login.model.LoginBean;
+
+/**
+ * Servlet implementation class LoginServlet
+ */
+
+@WebServlet("/login")
+public class LoginController extends HttpServlet {
+	private static final long serialVersionUID = 1L;
+	private LoginDao loginDao;
+
+	public void init() {
+		loginDao = new LoginDao();
+	}
+
+	protected void doGet(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
+		response.sendRedirect("login/login.jsp");
+	}
+
+	protected void doPost(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
+		authenticate(request, response);
+	}
+
+	private void authenticate(HttpServletRequest request, HttpServletResponse response)
+			throws IOException, ServletException {
+		String username = request.getParameter("username");
+		String password = request.getParameter("password");
+		LoginBean loginBean = new LoginBean();
+		loginBean.setUsername(username);
+		loginBean.setPassword(password);
+
+		try {
+			if (loginDao.validate(loginBean)) {
+				RequestDispatcher dispatcher = request.getRequestDispatcher("classes/classes-list.jsp");
+				dispatcher.forward(request, response);
+			} else {
+				HttpSession session = request.getSession();
+				// session.setAttribute("user", username);
+				// response.sendRedirect("login.jsp");
+			}
+		} catch (ClassNotFoundException e) {
+			e.printStackTrace();
+		}
+
+	}
+}
