--- src/main/java/net/gym/login/controller/LoginServlet.java
@@ -1,49 +1,0 @@
-package net.gym.login.controller;
-
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
-import net.gym.login.model.Login;
-import net.gym.login.dao.LoginDao;
-
-/**
- * Servlet implementation class LoginServlet
- */
-@WebServlet("/login")
-public class LoginServlet extends HttpServlet {
-    private static final long serialVersionUID = 1L;
-    private LoginDao loginDao;
-
-    public void init() {
-        loginDao = new LoginDao();
-    }
-
-    protected void doPost(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
-
-        String username = request.getParameter("username");
-        String password = request.getParameter("password");
-        Login login = new Login();
-        login.setUsername(username);
-        login.setPassword(password);
-
-        try {
-            if (loginDao.validate(login)) {
-                //HttpSession session = request.getSession();
-                // session.setAttribute("username",username);
-                response.sendRedirect("loginsuccess.jsp");
-            } else {
-                HttpSession session = request.getSession();
-                //session.setAttribute("user", username);
-                //response.sendRedirect("login.jsp");
-            }
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        }
-    }
-}
