+++ src/main/java/net/gym/login/model/LoginBean.java
@@ -1,0 +1,25 @@
+package net.gym.login.model;
+
+import java.io.Serializable;
+
+public class LoginBean implements Serializable {
+	private static final long serialVersionUID = 1L;
+	private String username;
+	private String password;
+
+	public String getUsername() {
+		return username;
+	}
+
+	public void setUsername(String username) {
+		this.username = username;
+	}
+
+	public String getPassword() {
+		return password;
+	}
+
+	public void setPassword(String password) {
+		this.password = password;
+	}
+}
